requirements.txt
customtkinter==5.1.3
PyAudio==0.2.14
PyAudioWPatch==0.2.12.7
faster-whisper~=1.1.0
Wave==0.0.2
google-genai
openai==1.61.1
openai-whisper==20230314
numpy==1.24.3
--extra-index-url https://download.pytorch.org/whl/cu126
torch~=2.5.1+cu121
torchaudio==2.6.0+cu126
python-dotenv~=1.0.1
deepgram-sdk